+++ hydromt/drivers/zarr_driver.py
@@ -9,8 +9,8 @@ from pyproj import CRS
 
 from hydromt._typing import Bbox, Geom
 from hydromt._typing.error import NoDataStrategy
-from hydromt.driver.preprocessing import PREPROCESSORS
-from hydromt.driver.rasterdataset_driver import RasterDatasetDriver
+from hydromt.drivers.preprocessing import PREPROCESSORS
+from hydromt.drivers.rasterdataset_driver import RasterDatasetDriver
 
 
 class ZarrDriver(RasterDatasetDriver):
@@ -43,10 +43,11 @@ class ZarrDriver(RasterDatasetDriver):
             uri,
             self.filesystem,
             bbox=bbox,
-            geom=mask,
+            mask=mask,
             buffer=buffer,
             predicate=predicate,
             variables=variables,
+            zoom_level=zoom_level,
             handle_nodata=handle_nodata,
             **kwargs,
         )