+++ hydromt/metadata_resolver/convention_resolver.py
@@ -3,15 +3,16 @@
 from functools import reduce
 from itertools import product
 from logging import Logger, getLogger
+from re import compile as compile_regex
+from re import error as regex_error
 from string import Formatter
-from typing import Any, Dict, Iterable, List, Optional, Set, Tuple
+from typing import Any, Dict, Iterable, List, Optional, Pattern, Set, Tuple
 
-import geopandas as gpd
 import numpy as np
 import pandas as pd
 from fsspec import AbstractFileSystem
 
-from hydromt._typing import Bbox, NoDataStrategy, Predicate, TimeRange
+from hydromt._typing import Bbox, Geom, NoDataStrategy, Predicate, TimeRange
 
 from .metadata_resolver import MetaDataResolver
 
@@ -21,28 +22,33 @@ logger: Logger = getLogger(__name__)
 class ConventionResolver(MetaDataResolver):
     """MetaDataResolver using HydroMT naming conventions."""
 
-    _uri_placeholders = frozenset({"year", "month", "variable", "zoom_level"})
+    _uri_placeholders = frozenset({"year", "month", "variable", "zoom_level", "name"})
 
     def _expand_uri_placeholders(
         self,
         uri: str,
         time_tuple: Optional[Tuple[str, str]] = None,
         variables: Optional[Set[str]] = None,
-    ) -> Tuple[str, List[str]]:
+    ) -> Tuple[str, List[str], Pattern[str]]:
         """Expand known placeholders in the URI."""
         keys: list[str] = []
+        pattern: str = ""
 
         if "{" in uri:
             uri_expanded = ""
             for literal_text, key, fmt, _ in Formatter().parse(uri):
                 uri_expanded += literal_text
+                pattern += literal_text
                 if key is None:
                     continue
+                pattern += "(.*)"
                 key_str = "{" + f"{key}:{fmt}" + "}" if fmt else "{" + key + "}"
                 # remove unused fields
                 if key in ["year", "month"] and time_tuple is None:
                     uri_expanded += "*"
                 elif key == "variable" and variables is None:
+                    uri_expanded += "*"
+                elif key == "name":
                     uri_expanded += "*"
                 # escape unknown fields
                 elif key is not None and key not in self._uri_placeholders:
@@ -52,7 +58,14 @@ class ConventionResolver(MetaDataResolver):
                     keys.append(key)
             uri = uri_expanded
 
-        return (uri, keys)
+        # darn windows paths creating invalid escape sequences grrrrr
+        try:
+            regex = compile_regex(pattern)
+        except regex_error:
+            # try it as raw path if regular string fails
+            regex = compile_regex(pattern.encode("unicode_escape").decode())
+
+        return (uri, keys, regex)
 
     def _get_dates(
         self,
@@ -83,8 +96,8 @@ class ConventionResolver(MetaDataResolver):
         *,
         timerange: Optional[TimeRange] = None,
         bbox: Optional[Bbox] = None,
-        # TODO: align? -> from RasterDatasetAdapter
-        mask: Optional[gpd.GeoDataFrame] = None,
+        # TODO: align? https://github.com/Deltares/hydromt/issues/874
+        mask: Optional[Geom] = None,
         buffer: float = 0.0,
         predicate: Predicate = "intersects",
         variables: Optional[List[str]] = None,
@@ -94,7 +107,7 @@ class ConventionResolver(MetaDataResolver):
         **kwargs,
     ) -> List[str]:
         """Resolve the placeholders in the URI."""
-        uri_expanded, keys = self._expand_uri_placeholders(uri, timerange, variables)
+        uri_expanded, keys, _ = self._expand_uri_placeholders(uri, timerange, variables)
         if timerange:
             dates = self._get_dates(keys, timerange)
         else: