+++ hydromt/plugins.py
@@ -1,11 +1,13 @@
 """Implementation of the mechanism to access the plugin entrypoints."""
 
+import inspect
 from typing import TYPE_CHECKING, Dict, Optional, Type, TypedDict, cast
 
 from importlib_metadata import entry_points
 
 if TYPE_CHECKING:
     from hydromt.components import ModelComponent  # noqa
+    from hydromt.drivers import BaseDriver
     from hydromt.models import Model  # noqa
 
 __all__ = ["PLUGINS"]
@@ -44,13 +46,15 @@ def _format_metadata(metadata: Dict[str, str]) -> str:
 class Plugins:
     """The model catalogue provides access to plugins."""
 
-    def __init__(self):
+    def __init__(self) -> None:
         """Initiate the catalog object."""
         self._component_plugins: Optional[Dict[str, Plugin]] = None
+        self._driver_plugins: Optional[Dict[str, Plugin]] = None
         self._model_plugins: Optional[Dict[str, Plugin]] = None
 
     def _initialize_plugins(self) -> None:
         self._component_plugins = _discover_plugins(group="hydromt.components")
+        self._driver_plugins = _discover_plugins(group="hydromt.drivers")
         self._model_plugins = _discover_plugins(group="hydromt.models")
 
     @property
@@ -70,6 +74,22 @@ class Plugins:
                     for name, value in self._component_plugins.items()
                 },
             )
+
+    @property
+    def driver_plugins(self) -> dict[str, Type["BaseDriver"]]:
+        """Load and provide access to all known driver plugins."""
+        if self._driver_plugins is None:
+            self._initialize_plugins()
+
+        if self._driver_plugins is None:
+            # core itself exposes plugins so if we can't find anything, something is wrong
+            raise RuntimeError("Could not load any driver plugins")
+
+        return {
+            name: value["type"]
+            for name, value in self._driver_plugins.items()
+            if not inspect.isabstract(value["type"])
+        }
 
     @property
     def model_plugins(self) -> dict[str, type["Model"]]:
@@ -116,6 +136,25 @@ class Plugins:
                 {k: v for k, v in self._component_plugins.items() if k != "type"},
             )
 
+    @property
+    def driver_metadata(self) -> Dict[str, Dict[str, str]]:
+        """Load and provide access to all known driver plugin metadata."""
+        if self._driver_plugins is None:
+            self._initialize_plugins()
+
+        if self._driver_plugins is None:
+            # core itself exposes plugins so if we can't find anything, something is wrong
+            raise RuntimeError("Could not load any driver plugins")
+        else:
+            return cast(
+                Dict[str, Dict[str, str]],
+                {
+                    k: v
+                    for k, v in self._driver_plugins.items()
+                    if not inspect.isabstract(v["type"])  # filter ABCs from core
+                },
+            )
+
     def model_summary(self) -> str:
         """Generate string representation containing the registered model entrypoints."""
         s = ""
@@ -125,6 +164,14 @@ class Plugins:
             map(_format_metadata, self.model_metadata.values())
         )
         return f"Model plugins:\n\t- {model_plugins}"
+
+    def driver_summary(self) -> str:
+        """Generate string representation container the registered driver entrypoints."""
+        self._initialize_plugins()
+        driver_plugins = "\n\t ".join(
+            list(map(_format_metadata, self.driver_metadata.values()))
+        )
+        return f"Driver Plugins:\n\t- {driver_plugins}"
 
     def component_summary(self) -> str:
         """Generate string representation containing the registered component entrypoints."""