+++ tests/metadata_resolvers/test_convention_resolver.py
@@ -64,3 +64,13 @@ class TestConventionResolver:
         uri = "/{unknown_key}_0_precip_2020_01.nc"
         resolver = ConventionResolver()
         assert resolver.resolve(uri, test_filesystem) == [uri]
+
+    def test_capture_regex(self):
+        pat = "here-is-some-more-leading-{year}-text-for-{month}-you-{variable}.pq"
+        example = "here-is-some-more-leading-2024-04-02-text-for-era5-you-0001.pq"
+        resolver = ConventionResolver()
+
+        glob, keys, regex = resolver._expand_uri_placeholders(pat)
+        assert glob == "here-is-some-more-leading-*-text-for-*-you-*.pq"
+        # we know regex will match, so type ignore is safe
+        assert regex.match(example).groups() == ("2024-04-02", "era5", "0001")  # type: ignore